Add a Godot 4.x starter project for Squash The Creeps 3D
The docs currently referencing downloading an asset pack wrapped in a
3.x project. That is now part of an archived repository. The demo
project was migrated as a 4.x project to, but it is the finished project
instead of the assets and a minimum project like the getting started
tutorial asks for.

The files added were copied from the current 4.x finished demo project.

I followed the entire 3d tutorial, and I am making a PR to the docs
repository to update a links to this starter projects and a few changes
based on 4.x updates. I will link the two PRs as a reference.

Current Docs:
* https://docs.godotengine.org/en/stable/getting_started/first_3d_game/index.html

Old starter repository:
https://github.com/godotengine/godot-3d-dodge-the-creeps

Current demo project:
https://github.com/godotengine/godot-demo-projects/tree/master/3d/squash_the_creeps